backend_IA/requirements.txt
fastapi
uvicorn
numpy
pillow
opencv-python
torch
torchvision
timm
git+https://github.com/isl-org/MiDaS.git@v3_1
